Update API and integrate with standalone C library
This is a significant overhaul of the CAN library.  It updates
the API to make it easier to use and more flexible to add
future capabilities.  The API has been changed so that a
CanMessage is now an object that contains all the parts of
a message: an ID, an extended flag, a data array and the
length of the data.  The Can class is used to configure CAN
and receive messages.  The CanMessage class is used to
create, send, and parse messages.

The file structure has also been updated so that the MCP2515.cpp
and spi.cpp files can be taken and used in non-arduino projects.
They are named as C++ files so that Arduino will compile them,
but are written as standard C, and can be treated as such simply
by changing the extension to ".c".  If you use the Arduino as a
development platform without the Arduino software, or have
developed an arduino program and want to move it to plain AVR,
standalone libraries will make it easier.