Specify CAN speed based on bit-width
Specify the CAN speed by passing the desired CAN bit-width to
the begin/init functions.  Instead of being hard-coded, the
segment widths are calculated based on the desired bit width.
This allows virtually any frequency that the MCP2515 is capable
of using to be specified, instead of just a preconfigured few.