Allow multiple data values per message
This change allows a CAN message to be packed with more than
one variable.  Any of the set functions can be called to fill
the next available bytes in a message with a given value.
When parsing a message, the data must be removed in the same
order in which they were added.