CAN: Add explicit filters for both std and extd IDs
The mask registers on the MCP2515 do not contain a mask bit for the
extended bit in the filter register, meaning that each filter is for
either a standard or extended message.  In the default configuration, no
filters are accepting extended messages.

This change adds explicit filters to accept both standard and extended
messages.  All mask bits are set to zero so that all messages of both
types are accepted.